[refactor] index에서 렌더링 하던것을 각 class에서 담당  #25
