+++ FE/src/js/index.js
@@ -1,6 +1,6 @@
 import { tplHeader } from "./tpl/tplHeader.js";
 import { _$, fetchGetData } from "./lib/util.js";
-import { CardView } from "./card/cardView.js";
+import { Card } from "./card/card.js";
 import { Column } from "./column/column.js";
 import { AddCard } from "./column/addCard.js";
 import { mock } from "./mock.js";
@@ -13,33 +13,21 @@ function init() {
 
   const addCard = new AddCard();
   const column = new Column(addCard);
-  const cardView = new CardView();
+  const card = new Card();
   // const card = new Card(cardView);
-  fetchInitialData(cardView, column);
+  fetchInitialData(card, column);
 }
 
-function fetchInitialData(cardView, column) {
+function fetchInitialData(card, column) {
   const url =
     "https://cors-anywhere.herokuapp.com/http://15.165.163.174:8080/mock";
 
-  // fetchGetData(url).then((initialData) =>
-  //   handleInitialData(initialData.categories, cardView, column)
-  // );
-
-  handleInitialData(mock.categories, cardView, column);
-}
-
-function handleInitialData(initialData, cardView, column) {
-  initialData.forEach((columnData) => {
-    const { id, name, cards } = columnData;
-
-    column.columnRender(id);
-    column.columnNameRender(id, name);
-    cards.forEach((card) => cardView.cardRender(id, card));
-    column.numberOfCardsRender(id);
+  fetchGetData(url).then((initialData) => {
+    column.init(initialData.categories);
+    card.init(initialData.categories);
   });
-
-  column.addColumnRender();
+  // column.init(mock.categories);
+  // card.init(mock.categories);
 }
 
 window.addEventListener("DOMContentLoaded", init);
